Remove all external dependencies
doctest brought into the project proper, since it is required in order just to
use this project as a library. It promises its cruft is removed if not actually
compiling the project to test it.

Also, I've settled on making the library generic, so any knowledge about a
specific rational type (boost::rational being the one relevant to current
discussion) is unnecessary. I'm likely to include an optional, simple rational
type in the library in the future.