Move all unit test code out of library proper.
For one, I couldn't figure out how to not pollute the macro space for users of
this project who also wanted to use doctest. Also, even with doctest's claims
of not hampering compile times or increasing the size of binaries, it's still
controversial to add *anything* to the final product that doesn't directly
benefit the library user. If the project is committed at all, the tests
(theoretically) have passed.